ci: Add OpenSSF scorecard setup
This patch enables a new github action that runs on each commit and
updates the HedgeDoc OpenSSF scorecard score, which is a combination of
various project best practices that are actively worked on. This should
help to stay on top of current best practices and provide transparency
for users of HedgeDoc.

References:
https://github.com/ossf/scorecard
https://securityscorecards.dev/

Signed-off-by: Sheogorath <sheogorath@shivering-isles.com>