docs(SECURITY): Shift vulnerability reporting directly to GitHub
This patch adjust the SECURITY.md to follow a new workflow, which
results in reports ending up on GitHub without going through my mailbox,
this frees resources on my side as well as reducing the bus factor.

Since most of the time, I do an editorial copy of the content of the
E-Mail into GitHub's vulnerability template, this should help to reduce
manual processes.

Signed-off-by: Sheogorath <sheogorath@shivering-isles.com>