fix: use github actions artifact to provide frontend e2e to cypress containers
something in our CI setup lets turbo believe that the environment has
changed so much that the build can't be restored from the cache.
Therefore all the cypress jobs are rebuilding the frontend on their own.
It probably has something to do with the fact that the cypress jobs are not
running with the ubuntu image but with the cypress docker image.
I couldn't find out how to fix this, so I changed the CI to use a GitHub
actions artifact.

Signed-off-by: Tilman Vatteroth <git@tilmanvatteroth.de>