chore(deps): use the same ts-node version in frontend/backend
This changes the specified version from main branch to a specific commit.
Hopefully this changes will result in (better) caching of the ts-node
package, especially in the GitHub CI. By setting a specific commit,
the package is also pinned like all other dependencies and therefore
ensures a more predictionable behaviour.

Signed-off-by: Erik Michelson <github@erik.michelson.eu>