fix(docker): docker builds should build @hedgedoc/database as well
It's almost a wonder that everything worked previously, at least in
our already built dev environments. Fresh environments had no clue
that the database package needs to be build, because our turbo.json
had irritating (circular) dependencies on ^build, and the package.json
of the backend didn't even contain the database package as a dependency.

Signed-off-by: Erik Michelson <github@erik.michelson.eu>