fix(config): adjust media config type to Zod schema
Our config error messages are generated from the nested
Zod schema for the config. This means, if the path for
the type of media backend to use is
  mediaBackend.type
then the env variable should be
  HD_MEDIA_BACKEND_TYPE

Previously, we used some custom env variable names
like
  HD_MEDIA_BACKEND

This however resulted in the message that the env
variable HD_MEDIA_BACKEND is not set even if it was,
or the other way around, because the generated error
messages were not fitting the actual variable name.

Signed-off-by: Erik Michelson <github@erik.michelson.eu>